Combine BaseEditor and RemoteFileEditor
